services/members/functions/requirements.txt
firebase-admin==6.5.0
firebase-functions==0.4.1
PyJWT[crypto]==2.8.0
requests==2.31.0
google-cloud-storage==2.18.2
google-cloud-secret-manager==2.16.4
python-dotenv==1.0.1

# Testing dependencies
pytest==8.3.3
pytest-cov==6.0.0
